Moved astar into its own package
