+++ src/main/java/org/opentripplanner/visualizer/GraphVisualizer.java
@@ -3,10 +3,9 @@ package org.opentripplanner.visualizer;
 import com.google.common.collect.Sets;
 import org.locationtech.jts.geom.Coordinate;
 import javassist.Modifier;
-import org.opentripplanner.common.model.GenericLocation;
 import org.opentripplanner.graph_builder.annotation.GraphBuilderAnnotation;
 import org.opentripplanner.graph_builder.annotation.StopUnlinked;
-import org.opentripplanner.routing.algorithm.TraverseVisitor;
+import org.opentripplanner.routing.algorithm.astar.TraverseVisitor;
 import org.opentripplanner.routing.core.OptimizeType;
 import org.opentripplanner.routing.core.RoutingRequest;
 import org.opentripplanner.routing.core.State;
