+++ src/test/java/org/opentripplanner/graph_builder/module/osm/TestUnroutable.java
@@ -4,7 +4,7 @@ package org.opentripplanner.graph_builder.module.osm;
 import com.google.common.collect.Maps;
 import junit.framework.TestCase;
 import org.opentripplanner.openstreetmap.impl.AnyFileBasedOpenStreetMapProviderImpl;
-import org.opentripplanner.routing.algorithm.AStar;
+import org.opentripplanner.routing.algorithm.astar.AStar;
 import org.opentripplanner.routing.core.RoutingRequest;
 import org.opentripplanner.routing.core.TraverseMode;
 import org.opentripplanner.routing.graph.Edge;
