+++ src/test/java/org/opentripplanner/routing/algorithm/TestBanning.java
@@ -3,9 +3,9 @@ package org.opentripplanner.routing.algorithm;
 import junit.framework.TestCase;
 import org.junit.Ignore;
 import org.opentripplanner.model.FeedScopedId;
-import org.opentripplanner.model.Route;
 import org.opentripplanner.ConstantsForTests;
 import org.opentripplanner.common.model.T2;
+import org.opentripplanner.routing.algorithm.astar.AStar;
 import org.opentripplanner.routing.core.RoutingRequest;
 import org.opentripplanner.routing.core.State;
 import org.opentripplanner.routing.graph.Graph;
