+++ src/test/java/org/opentripplanner/routing/algorithm/TestGraphPath.java
@@ -9,6 +9,7 @@ import org.junit.Ignore;
 import org.opentripplanner.model.calendar.CalendarServiceData;
 import org.opentripplanner.ConstantsForTests;
 import org.opentripplanner.gtfs.GtfsContext;
+import org.opentripplanner.routing.algorithm.astar.AStar;
 import org.opentripplanner.routing.core.RoutingRequest;
 import org.opentripplanner.routing.core.State;
 import org.opentripplanner.routing.edgetype.factory.PatternHopFactory;
